+++ 2Homework25.02.22/StackCalculator/Solution/IStack.cs
@@ -1,0 +1,13 @@
+namespace StackCalculator.Solution;
+
+interface IStack
+{
+    void Push(float number);
+
+    float? Pop();
+    
+    float? Top { get; }
+    
+    bool IsEmpty { get; }
+
+}
