+++ 2Homework25.02.22/StackCalculator/Solution/Program.cs
@@ -1,0 +1,12 @@
+﻿namespace StackCalculator.Solution;
+
+
+internal class Program
+{
+    public static void Main(string[] args)
+    {
+        var x = new StackCalculator();
+        x.ChangeStack(StackCalculator.TypesOfStacks.ListStack);
+    }
+}
+
